Updates to rename class, take care of some comments
* rename S3AccessFromConfig -> AmazonS3ClientFromConfig
* change return type of list functions from List[*] to Seq[*]
* general cleanup
