+++ src/main/resources/reference.conf
@@ -7,7 +7,4 @@ c3p0 {
   jdbcUrl = "jdbc:postgresql://localhost:5432/decima"
   user = "blist"
   password = "blist"
-#  minPoolSize = 1
-#  acquireIncrement = 1
-#  maxPoolSize = 50
 }
