+++ src/main/scala/com/socrata/decima/data_access/S3Access.scala
@@ -1,7 +1,7 @@
 package com.socrata.decima.data_access
 
 import com.amazonaws.services.s3._
-import com.amazonaws.services.s3.model.ListObjectsRequest
+import com.amazonaws.services.s3.model.{ListObjectsRequest, ObjectListing}
 import com.socrata.decima.models.S3BuildInfo
 import grizzled.slf4j.Logging
 import org.yaml.snakeyaml.Yaml
@@ -10,7 +10,7 @@ import org.yaml.snakeyaml.constructor.Constructor
 import scala.collection.JavaConverters._
 
 trait S3AccessBase {
-  def listBuildPaths(project: String): List[String]
+  def listBuildPaths(project: String): Seq[String]
   def getBuildInfo(project: String, buildId: String): S3BuildInfo
 }
 
@@ -25,18 +25,22 @@ class S3Access(s3: AmazonS3Client, bucketName: String) extends S3AccessBase with
    * @param project: the prefix of the project to use in finding the build
    * @return a list of build prefixes, e.g. ["CoreServer/12341234_12", ...]
    */
-  override def listBuildPaths(project: String): List[String] = {
+  override def listBuildPaths(project: String): Seq[String] = {
     var objectListing = s3.listObjects(new ListObjectsRequest()
                                             .withBucketName(bucketName)
                                             .withPrefix(s"$project/")
                                             .withDelimiter("/"))
-    var prefixes = objectListing.getCommonPrefixes.asScala.toList
+    var prefixes = getBuildPrefixes(objectListing)
     while(objectListing.isTruncated) {
       objectListing = s3.listNextBatchOfObjects(objectListing)
-      prefixes ++= objectListing.getCommonPrefixes.asScala.toList
+      prefixes ++= getBuildPrefixes(objectListing)
     }
     if (prefixes.isEmpty) throw new RuntimeException(s"Unable to find any builds under $project")
     prefixes
+  }
+
+  private def getBuildPrefixes(objectListing: ObjectListing): Seq[String] = {
+    objectListing.getCommonPrefixes.asScala
   }
 
   /**
@@ -54,7 +58,7 @@ class S3Access(s3: AmazonS3Client, bucketName: String) extends S3AccessBase with
       val inputStream = buildInfoObject.getObjectContent
       val buildInfo = yaml.load(inputStream).asInstanceOf[S3BuildInfo]
       buildInfoObject.close()
-    buildInfo.s3Url = s3Url
+      buildInfo.s3Url = s3Url
       buildInfo
     } catch {
       case e: Exception =>
